pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ashare-quant"
version = "0.1.0"
description = "A-share quant analysis MVP"
requires-python = ">=3.11"
dependencies = [
  "akshare>=1.14.70",
  "baostock>=0.8.9",
  "efinance>=0.5.5.2",
  "numpy>=2.2.0",
  "pandas>=2.2.3",
  "plotly>=6.0.0",
  "PyYAML>=6.0.2",
  "streamlit>=1.44.0",
  "tushare>=1.4.24",
]

[tool.setuptools.packages.find]
where = ["src"]
